+++ iris/transport_metrics_test.go
@@ -1,0 +1,6 @@
+package iris_test
+
+import "github.com/park285/iris-client-go/iris"
+
+var _ iris.TransportMetrics = iris.NoopTransportMetrics{}
+var _ iris.ClientOption = iris.WithTransportMetrics(iris.NoopTransportMetrics{})